requirements.txt
blitz-bayesian-pytorch==0.2.7
psutil
box2d-py==2.3.8
pyglet==1.5.0
wandb==0.10.29
memory_profiler
pytorch-memlab==0.2.3
ipython==7.23.1
matplotlib==3.4.2
termcolor
ipywidgets
tqdm
nvidia-ml-py3
torchbnn
highway-env